move gs generators into src folder
