+++ packages/gluestick/src/generator/templates/500hbs.js
