+++ packages/gluestick/src/generator/templates/App.js
