+++ packages/gluestick/src/generator/templates/ContainerHomeTest.js
