+++ packages/gluestick/src/generator/templates/Dockerfile.js
