+++ packages/gluestick/src/generator/templates/Empty.js
