+++ packages/gluestick/src/generator/templates/EntryWrapper.js
