+++ packages/gluestick/src/generator/templates/Home.js
