+++ packages/gluestick/src/generator/templates/HomeApp.js
