+++ packages/gluestick/src/generator/templates/Index.js
