+++ packages/gluestick/src/generator/templates/InitBrowser.js
