+++ packages/gluestick/src/generator/templates/MasterLayout.js
