+++ packages/gluestick/src/generator/templates/MasterLayoutTest.js
