+++ packages/gluestick/src/generator/templates/Reducer.js
