+++ packages/gluestick/src/generator/templates/ReduxMiddleware.js
