+++ packages/gluestick/src/generator/templates/WebpackAdditions.js
