+++ packages/gluestick/src/generator/templates/_eslintrc.js
@@ -1,5 +1,5 @@
 /* @flow */
-import type { CreateTemplate } from '../../src/types';
+import type { CreateTemplate } from '../../types';
 
 module.exports = (createTemplate: CreateTemplate) => createTemplate`
 {
