+++ packages/gluestick/src/generator/templates/_flowconfig.js
