+++ packages/gluestick/src/generator/templates/babelrc.js
