+++ packages/gluestick/src/generator/templates/caching.server.js
