+++ packages/gluestick/src/generator/templates/dockerignore.js
