+++ packages/gluestick/src/generator/templates/entries.js
