+++ packages/gluestick/src/generator/templates/gitignore.js
