+++ packages/gluestick/src/generator/templates/gluestick.config.js
