+++ packages/gluestick/src/generator/templates/gluestick.hooks.js
