+++ packages/gluestick/src/generator/templates/gluestick.plugins.js
