+++ packages/gluestick/src/generator/templates/webpack.hooks.js
